requirements.txt
# متطلبات البوت الأساسية
python-telegram-bot==20.7
telethon==1.34.0

# قاعدة البيانات
sqlite3
sqlalchemy==2.0.23
alembic==1.13.1

# معالجة النصوص والوقت
emoji==2.8.0
python-dateutil==2.8.2
pytz==2023.3

# الشبكة والأمان
aiohttp==3.9.1
cryptography==41.0.8
requests==2.31.0

# معالجة البيانات
pandas==2.1.4
numpy==1.25.2

# السجلات والمراقبة
loguru==0.7.2

# البيئة والإعدادات
python-dotenv==1.0.0
pydantic==2.5.2

# اختبارات
pytest==7.4.3
pytest-asyncio==0.21.1

# إنتاج
gunicorn==21.2.0
uvicorn==0.24.0

# أدوات إضافية
pillow==10.1.0
qrcode==7.4.2
